deploy fixes
